X00TD: Generate non-hlos image mount points for A/B builds
On builds where A/B support is enabled the system image acts as the
rootfs. In this case we can no longer create the non-hlos image mount
points at run time as we used to. We now create them as part of the
build itself. As for the non A/B cases it does not hurt having them
created at build time and simplifies selinux rules.

Change-Id: I30d98f7c50fca3ce7117b9b8d87150655bb5b440
Signed-off-by: Navin136 <nkwhitehat@gmail.com>